working RMS readout

+++ hit2023v2/display.h
@@ -51,9 +51,9 @@ private:
     QMap<QString, QVector<unsigned short>> backgroundDataMap; // Map to store background data for each plane
     bool subtractBackground = false; // Flag to track if background subtraction is enabled
 
-    QMap<QString, QVector<unsigned short>> calibrationDataMap; // Map to store calibration data for each plane
+    QMap<QString, QVector<float>> calibrationDataMap; // Map to store calibration data for each plane
     bool applyCalibration = false; // Flag to track if calibration should be applied
-    QVector<unsigned short> calibrationData; // Stores the loaded calibration data
+    QVector<float> calibrationData; // Stores the loaded calibration data
     QCheckBox *checkBoxExpertMode; // Expert Mode checkbox
     bool expertModeEnabled = false; // Flag to track if expert mode is enabled
 };
